Hide the base implementation of functions behind an interface
If someone, for some reason, would like to implement functions from scratch.
