+++ jmespath-core/src/main/java/io/burt/jmespath/function/MapFunction.java
@@ -7,7 +7,7 @@ import io.burt.jmespath.Adapter;
 import io.burt.jmespath.JmesPathType;
 import io.burt.jmespath.node.Node;
 
-public class MapFunction extends Function {
+public class MapFunction extends BaseFunction {
   public MapFunction() {
     super(
       ArgumentConstraints.expression(),
