+++ jmespath-core/src/main/java/io/burt/jmespath/function/ReverseFunction.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import io.burt.jmespath.Adapter;
 import io.burt.jmespath.JmesPathType;
 
-public class ReverseFunction extends Function {
+public class ReverseFunction extends BaseFunction {
   public ReverseFunction() {
     super(ArgumentConstraints.typeOf(JmesPathType.ARRAY, JmesPathType.STRING));
   }
