+++ jmespath-core/src/main/java/io/burt/jmespath/function/SortFunction.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import io.burt.jmespath.Adapter;
 import io.burt.jmespath.JmesPathType;
 
-public class SortFunction extends Function {
+public class SortFunction extends BaseFunction {
   public SortFunction() {
     super(
       ArgumentConstraints.arrayOf(
