+++ jmespath-core/src/main/java/io/burt/jmespath/function/ToArrayFunction.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 import io.burt.jmespath.Adapter;
 import io.burt.jmespath.JmesPathType;
 
-public class ToArrayFunction extends Function {
+public class ToArrayFunction extends BaseFunction {
   public ToArrayFunction() {
     super(ArgumentConstraints.anyValue());
   }
