+++ jmespath-core/src/main/java/io/burt/jmespath/function/ValuesFunction.java
@@ -5,7 +5,7 @@ import java.util.List;
 import io.burt.jmespath.Adapter;
 import io.burt.jmespath.JmesPathType;
 
-public class ValuesFunction extends Function {
+public class ValuesFunction extends BaseFunction {
   public ValuesFunction() {
     super(ArgumentConstraints.typeOf(JmesPathType.OBJECT));
   }
